move some util classes to iot-service-base
