+++ iot/iot-device-registry-infinispan/src/main/java/io/enmasse/iot/registry/infinispan/device/AbstractCredentialsManagementService.java
@@ -5,7 +5,7 @@
 
 package io.enmasse.iot.registry.infinispan.device;
 
-import static io.enmasse.iot.registry.infinispan.device.data.DeviceKey.deviceKey;
+import static io.enmasse.iot.service.base.infinispan.device.DeviceKey.deviceKey;
 import static io.enmasse.iot.service.base.utils.MoreFutures.completeHandler;
 import static java.net.HttpURLConnection.HTTP_NOT_FOUND;
 
@@ -19,11 +19,11 @@ import org.eclipse.hono.service.management.credentials.CredentialsManagementServ
 import org.infinispan.client.hotrod.RemoteCache;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import io.enmasse.iot.registry.infinispan.cache.AdapterCredentialsCacheProvider;
-import io.enmasse.iot.registry.infinispan.cache.DeviceManagementCacheProvider;
-import io.enmasse.iot.registry.infinispan.device.data.CredentialKey;
-import io.enmasse.iot.registry.infinispan.device.data.DeviceInformation;
-import io.enmasse.iot.registry.infinispan.device.data.DeviceKey;
+import io.enmasse.iot.service.base.infinispan.cache.AdapterCredentialsCacheProvider;
+import io.enmasse.iot.service.base.infinispan.cache.DeviceManagementCacheProvider;
+import io.enmasse.iot.service.base.infinispan.device.CredentialKey;
+import io.enmasse.iot.service.base.infinispan.device.DeviceInformation;
+import io.enmasse.iot.service.base.infinispan.device.DeviceKey;
 import io.enmasse.iot.registry.infinispan.tenant.TenantInformationService;
 import io.opentracing.Span;
 import io.vertx.core.AsyncResult;
