+++ iot/iot-device-registry-infinispan/src/main/java/io/enmasse/iot/registry/infinispan/tenant/KubernetesTenantInformationService.java
@@ -5,7 +5,8 @@
 
 package io.enmasse.iot.registry.infinispan.tenant;
 
-import static io.enmasse.iot.registry.infinispan.tenant.TenantHandle.of;
+import io.enmasse.iot.service.base.infinispan.tenant.TenantHandle;
+import static io.enmasse.iot.service.base.infinispan.tenant.TenantHandle.of;
 import static java.net.HttpURLConnection.HTTP_INTERNAL_ERROR;
 import static java.util.concurrent.CompletableFuture.completedFuture;
 import static java.util.concurrent.CompletableFuture.failedFuture;
