+++ iot/iot-service-base/src/main/java/io/enmasse/iot/service/base/infinispan/cache/DeviceConnectionCacheProvider.java
@@ -3,7 +3,7 @@
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
 
-package io.enmasse.iot.registry.infinispan.cache;
+package io.enmasse.iot.service.base.infinispan.cache;
 
 import org.infinispan.client.hotrod.RemoteCache;
 import org.infinispan.configuration.cache.CacheMode;
@@ -11,14 +11,14 @@ import org.infinispan.configuration.cache.Index;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import io.enmasse.iot.registry.infinispan.config.InfinispanProperties;
-import io.enmasse.iot.registry.infinispan.device.data.CredentialKey;
+import io.enmasse.iot.service.base.infinispan.config.InfinispanProperties;
+import io.enmasse.iot.service.base.infinispan.devcon.DeviceConnectionKey;
 
 @Component
-public class AdapterCredentialsCacheProvider extends AbstractCacheProvider {
+public class DeviceConnectionCacheProvider extends AbstractCacheProvider {
 
     @Autowired
-    public AdapterCredentialsCacheProvider(final InfinispanProperties properties) throws Exception {
+    public DeviceConnectionCacheProvider(final InfinispanProperties properties) throws Exception {
         super(properties);
     }
 
@@ -36,10 +36,8 @@ public class AdapterCredentialsCacheProvider extends AbstractCacheProvider {
                 .build();
     }
 
-
-    public RemoteCache<CredentialKey, String> getAdapterCredentialsCache() {
-        return getOrCreateCache(properties.getAdapterCredentialsCacheName(), buildConfiguration());
+    public RemoteCache<DeviceConnectionKey, String> getDeviceStateCache() {
+        return getOrCreateCache(properties.getDeviceStatesCacheName(), buildConfiguration());
     }
-
 
 }
