+++ iot/iot-service-base/src/main/java/io/enmasse/iot/service/base/infinispan/config/InfinispanProperties.java
@@ -3,7 +3,7 @@
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
 
-package io.enmasse.iot.registry.infinispan.config;
+package io.enmasse.iot.service.base.infinispan.config;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
